Update README.md
This script could use a lot of tweaking and could certainly be a lot shorter. However the target audience (hopefully) don't expect much.
I've seen a lot of other Zoom automations that feel really complicated(needing to install pyautogui, taking screenshots, updating the class time table in excel etc.) and are more sensitive to errors in code. This script looked really simple to the 2 guys I've shown this. Hence I'll let it slide.